add post and comment to user. and config user/admin can create post/comment but user cannot delete/update . admin can delete/update

+++ src/main/java/com/springboot/blog/service/impl/CommentServiceImpl.java
@@ -2,14 +2,18 @@ package com.springboot.blog.service.impl;
 
 import com.springboot.blog.entity.Comment;
 import com.springboot.blog.entity.Post;
+import com.springboot.blog.entity.User;
 import com.springboot.blog.exception.BlogAPIException;
 import com.springboot.blog.exception.ResourceNotFoundException;
 import com.springboot.blog.payload.CommentDto;
 import com.springboot.blog.repository.CommentRepository;
 import com.springboot.blog.repository.PostRepository;
+import com.springboot.blog.repository.UserRepository;
+import com.springboot.blog.security.JwtTokenProvider;
 import com.springboot.blog.service.CommentService;
 import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,19 +22,32 @@ import java.util.List;
 public class CommentServiceImpl implements CommentService{
     private CommentRepository commentRepository;
     private PostRepository postRepository;
+    private UserRepository userRepository;
+    private JwtTokenProvider jwtTokenProvider;
     private ModelMapper mapper;
-    public CommentServiceImpl(CommentRepository commentRepository, PostRepository postRepository, ModelMapper mapper) {
+    public CommentServiceImpl(CommentRepository commentRepository,
+                              PostRepository postRepository,
+                              UserRepository userRepository,
+                              ModelMapper mapper,
+                              JwtTokenProvider jwtTokenProvider) {
         this.commentRepository = commentRepository;
         this.postRepository = postRepository;
+        this.userRepository = userRepository;
+        this.jwtTokenProvider = jwtTokenProvider;
         this.mapper = mapper;
     }
     @Override
-    public CommentDto createComment(long postId,CommentDto commentDto) {
+    public CommentDto createComment(long postId,CommentDto commentDto,String token) {
         Comment newComment = mapToEntity(commentDto);
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new ResourceNotFoundException("Post", "id", postId)
         );
+        String username = jwtTokenProvider.getUsernameFromJWT(token);
+        User user = userRepository.findByUsername(username).orElseThrow(
+                () -> new UsernameNotFoundException("User not found with username. " + username)
+        );
         newComment.setPost(post);
+        newComment.setUser(user);
         Comment savedComment = commentRepository.save(newComment);
         return mapToDto(savedComment);
     }
@@ -70,7 +87,7 @@ public class CommentServiceImpl implements CommentService{
         return mapToDto(updatedComment);
     }
     @Override
-    public void deleteComment(long postId, long commentId) {
+    public void deleteComment(long postId, long commentId, String token) {
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new ResourceNotFoundException("Post", "id", postId)
         );
@@ -79,6 +96,10 @@ public class CommentServiceImpl implements CommentService{
         );
         if(!comment.getPost().getId().equals(post.getId())){
             throw new BlogAPIException(HttpStatus.BAD_REQUEST, "Comment does not belong to Post");
+        }
+        String username = jwtTokenProvider.getUsernameFromJWT(token);
+        if(!comment.getUser().getUsername().equals(username)){
+            throw new BlogAPIException(HttpStatus.BAD_REQUEST,"This is not your comment,Can not Delete");
         }
         commentRepository.delete(comment);
     }
